backend/requirements.txt
# Hermes Agent Backend Requirements

# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy==2.0.31
aiosqlite==0.20.0
alembic==1.13.2

# Data Validation
pydantic==2.8.2
pydantic-settings==2.4.0

# Task Scheduling
apscheduler==3.10.4

# HTTP Client
httpx==0.27.0

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Logging
loguru==0.7.2

# Retry Logic
tenacity==8.5.0

# JSON Processing
orjson==3.10.6
